+++ ui/tabs/src/TabsRoot.tsx
@@ -3,6 +3,7 @@ import * as TabsPrimitive from "@radix-ui/react-tabs";
 import { styled } from "@washingtonpost/wpds-theme";
 
 import type * as WPDS from "@washingtonpost/wpds-theme";
+import { TabsContext } from "./context";
 
 const StyledTabsRoot = styled(TabsPrimitive.Root, {});
 
@@ -17,11 +18,17 @@ export type TabsRootProps = {
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const TabsRoot = React.forwardRef<HTMLDivElement, TabsRootProps>(
-  ({ children, defaultValue = "", ...props }: TabsRootProps, ref) => {
+  ({ children, ...props }: TabsRootProps, ref) => {
     return (
-      <StyledTabsRoot defaultValue={defaultValue} {...props} ref={ref}>
-        {children}
-      </StyledTabsRoot>
+      <TabsContext.Provider
+        value={{
+          initialValue: props.defaultValue || props.value,
+        }}
+      >
+        <StyledTabsRoot {...props} ref={ref}>
+          {children}
+        </StyledTabsRoot>
+      </TabsContext.Provider>
     );
   }
 );