+++ ui/button/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.2](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.5.1...v1.5.2) (2023-05-02)
+
+**Note:** Version bump only for package @washingtonpost/wpds-button
+
+## [1.5.1](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.5.0...v1.5.1) (2023-05-02)
+
+### Bug Fixes
+
+- button density set to default value ([#381](https://github.com/washingtonpost/wpds-ui-kit/issues/381)) ([3ac5761](https://github.com/washingtonpost/wpds-ui-kit/commit/3ac57611e1573fcd5f5eeb4d5b2f48928492cb88))
+
 # [1.5.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v1.4.0...v1.5.0) (2023-04-18)
 
 **Note:** Version bump only for package @washingtonpost/wpds-button